src/budget_sync/requirements.txt
boto3==1.28.14
google-api-python-client==2.70.0
google-auth==2.20.0
google-auth-oauthlib==0.8.0
Flask==2.2.5
python-dotenv==1.0.0

# For development/testing purposes
pytest==7.3.1
python-dotenv

# Google libraries for the Budget Sync Lambda function
google-api-python-client>=2.0.0
google-cloud-bigquery>=3.0.0
google-auth>=2.0.0
google-auth-oauthlib>=0.4.0
